backend/requirements.txt
fastapi
uvicorn
requests
pandas
pymongo
python-dotenv
apscheduler
xgboost
scikit-learn
joblib